move revertToBlock to project actions context
