+++ frontend/src/pages/blocks/main/Main.tsx
@@ -14,30 +14,22 @@ import Table from "../../../components/table/Table";
 import { DecoratedPollingEntity } from "../../../hooks/use-timeout-polling";
 import { Block } from "@flowser/shared";
 import {
-  useGetCurrentProject,
   useGetPollingBlocks,
   useGetPollingEmulatorSnapshots,
 } from "../../../hooks/use-api";
-import toast from "react-hot-toast";
 import { SimpleButton } from "../../../components/simple-button/SimpleButton";
-import { ServiceRegistry } from "../../../services/service-registry";
-import { useErrorHandler } from "../../../hooks/use-error-handler";
 import { ReactComponent as SnapshotIcon } from "../../../assets/icons/snapshot.svg";
-import { useConfirmDialog } from "../../../contexts/confirm-dialog.context";
 import ReactTimeago from "react-timeago";
+import { useProjectActions } from "../../../contexts/project-actions.context";
 
 const columnHelper = createColumnHelper<DecoratedPollingEntity<Block>>();
 
 const Main: FunctionComponent = () => {
-  const { snapshotService } = ServiceRegistry.getInstance();
-  const { handleError } = useErrorHandler(Main.name);
   const { searchTerm, setPlaceholder, disableSearchBar } = useSearch();
   const { showNavigationDrawer } = useNavigation();
-  const { showDialog } = useConfirmDialog();
+  const { revertToBlock } = useProjectActions();
 
-  const { data: projectResponse } = useGetCurrentProject();
-  const { project } = projectResponse ?? {};
-  const { data: blocks, firstFetch, fetchAll } = useGetPollingBlocks();
+  const { data: blocks, firstFetch } = useGetPollingBlocks();
   const { data: emulatorSnapshots } = useGetPollingEmulatorSnapshots();
   const { filteredData } = useFilterData(blocks, searchTerm);
   const snapshotLookupByBlockId = useMemo(
@@ -47,39 +39,6 @@ const Main: FunctionComponent = () => {
       ),
     [emulatorSnapshots]
   );
-
-  async function onRevertToBlock(blockId: string) {
-    const isSnapshotEnabled = project?.emulator?.snapshot;
-    if (!isSnapshotEnabled) {
-      toast.error(
-        "Can't revert, because 'snapshot' option is not enabled in settings"
-      );
-      return;
-    }
-    const block = blocks.find((block) => block.id === blockId);
-    showDialog({
-      title: "Revert to snapshot",
-      body: (
-        <span style={{ textAlign: "center" }}>
-          Do you want to revert the emulator blockchain state to the block with
-          height <code>{block?.height}</code>?
-        </span>
-      ),
-      confirmBtnLabel: "REVERT",
-      cancelBtnLabel: "CANCEL",
-      onConfirm: async () => {
-        try {
-          const snapshot = await snapshotService.revertTo({
-            blockId,
-          });
-          fetchAll();
-          toast.success(`Reverted to "${snapshot.snapshot?.description}"`);
-        } catch (e) {
-          handleError(e);
-        }
-      },
-    });
-  }
 
   useEffect(() => {
     setPlaceholder("Search for block ids, parent ids, time, ...");
@@ -123,7 +82,7 @@ const Main: FunctionComponent = () => {
               {snapshot && (
                 <SimpleButton
                   style={{ marginLeft: 5 }}
-                  onClick={() => onRevertToBlock(block.id)}
+                  onClick={() => revertToBlock(block.id)}
                 >
                   <SnapshotIcon />
                 </SimpleButton>
