+++ frontend/src/pages/project/main/ProjectItemBadge.tsx
@@ -1,11 +1,21 @@
 import React, { FunctionComponent, HTMLAttributes } from "react";
 import classes from "./ProjectItemBadge.module.scss";
+import { SimpleButton } from "../../../components/simple-button/SimpleButton";
+import classNames from "classnames";
 
 export type BadgeProps = HTMLAttributes<HTMLSpanElement>;
 
 export const ProjectItemBadge: FunctionComponent<BadgeProps> = ({
   className,
   children,
+  ...restProps
 }) => {
-  return <span className={`${classes.root} ${className}`}>{children}</span>;
+  return (
+    <SimpleButton
+      className={classNames(classes.root, className)}
+      {...restProps}
+    >
+      {children}
+    </SimpleButton>
+  );
 };
